+++ gamestonk_terminal/stocks/fundamental_analysis/excel/variables.py
@@ -226,191 +226,4 @@ tickers = [
     "AER",
     "AERI",
     "AES",
-    "AESE",
-    "AEVA",
-    "AEY",
-    "AEYE",
-    "AEZS",
-    "AFAQ",
-    "AFBI",
-    "AFCG",
-    "AFG",
-    "AFI",
-    "AFIB",
-    "AFIN",
-    "AFL",
-    "AFMD",
-    "AFRM",
-    "AFYA",
-    "AG",
-    "AGAC",
-    "AGBA",
-    "AGC",
-    "AGCB",
-    "AGCO",
-    "AGE",
-    "AGEN",
-    "AGFS",
-    "AGFY",
-    "AGGR",
-    "AGI",
-    "AGIO",
-    "AGL",
-    "AGLE",
-    "AGM",
-    "AGM.A",
-    "AGMH",
-    "AGNC",
-    "AGO",
-    "AGR",
-    "AGRI",
-    "AGRO",
-    "AGRX",
-    "AGS",
-    "AGTC",
-    "AGTI",
-    "AGX",
-    "AGYS",
-    "AHAC",
-    "AHCO",
-    "AHH",
-    "AHPI",
-    "AHT",
-    "AI",
-    "AIG",
-    "AIH",
-    "AIHS",
-    "AIKI",
-    "AIM",
-    "AIMC",
-    "AIN",
-    "AINC",
-    "AINV",
-    "AIR",
-    "AIRC",
-    "AIRG",
-    "AIRI",
-    "AIRT",
-    "AIT",
-    "AIV",
-    "AIZ",
-    "AJAX",
-    "AJG",
-    "AJRD",
-    "AJX",
-    "AKAM",
-    "AKBA",
-    "AKIC",
-    "AKO.A",
-    "AKO.B",
-    "AKR",
-    "AKRO",
-    "AKTS",
-    "AKTX",
-    "AKU",
-    "AKUS",
-    "AKYA",
-    "AL",
-    "ALAC",
-    "ALB",
-    "ALBO",
-    "ALC",
-    "ALCC",
-    "ALCO",
-    "ALDX",
-    "ALE",
-    "ALEC",
-    "ALEX",
-    "ALF",
-    "ALG",
-    "ALGM",
-    "ALGN",
-    "ALGS",
-    "ALGT",
-    "ALHC",
-    "ALIM",
-    "ALIT",
-    "ALJJ",
-    "ALK",
-    "ALKS",
-    "ALKT",
-    "ALL",
-    "ALLO",
-    "ALLT",
-    "ALLY",
-    "ALNA",
-    "ALNY",
-    "ALPN",
-    "ALRM",
-    "ALRN",
-    "ALRS",
-    "ALSN",
-    "ALT",
-    "ALTA",
-    "ALTG",
-    "ALTM",
-    "ALTO",
-    "ALTR",
-    "ALTU",
-    "ALV",
-    "ALVR",
-    "ALX",
-    "ALXO",
-    "ALYA",
-    "ALZN",
-    "AM",
-    "AMAL",
-    "AMAM",
-    "AMAO",
-    "AMAT",
-    "AMBA",
-    "AMBC",
-    "AMBO",
-    "AMBP",
-    "AMC",
-    "AMCR",
-    "AMCX",
-    "AMD",
-    "AME",
-    "AMED",
-    "AMEH",
-    "AMG",
-    "AMGN",
-    "AMH",
-    "AMHC",
-    "AMK",
-    "AMKR",
-    "AMN",
-    "AMNB",
-    "AMOT",
-    "AMOV",
-    "AMP",
-    "AMPE",
-    "AMPG",
-    "AMPH",
-    "AMPI",
-    "AMPY",
-    "AMR",
-    "AMRC",
-    "AMRK",
-    "AMRN",
-    "AMRS",
-    "AMRX",
-    "AMS",
-    "AMSC",
-    "AMSF",
-    "AMST",
-    "AMSWA",
-    "AMT",
-    "AMTB",
-    "AMTBB",
-    "AMTI",
-    "AMTX",
-    "AMWD",
-    "AMWL",
-    "AMX",
-    "AMYT",
-    "AMZN",
-    "AN",
-    "ANAB",
 ]
