+++ pyproject.toml
@@ -80,6 +80,7 @@ sentiment-investor = "^2.0.0"
 fred = "^3.1"
 openpyxl = "^3.0.7"
 vcrpy = "^4.1.1"
+coverage = "^5.5"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.2"
